fix(capture): write measured session fps into annotation.json video.fps
The capture arms one frame per engine tick, so the session's true rate is
whatever PIE achieved (e.g. ~10 fps on a loaded editor). Encoding at a fixed
fps=30 played a 12s take back in 4s (3x fast-forward). Now the subsystem
stamps the world-time of the first/last ARMED frame (same clock as the
labels' "t") and writes the MEASURED rate at finalize:

  video.fps = (armed_frames - 1) / (t_last - t_first)   [clamped 1..240,
  rounded to 3 decimals; may be fractional; falls back to 30 when <2 frames]

The host encoder already reads video.fps, so mp4 duration now matches the
gameplay span and motion plays at natural speed. total_frames /
affected_frames / labels are untouched -- this is purely the time base.
Settle-phase gaps (uncaptured frames) smear into the average; accepted.

Reverses the fixed-fps=30 decision (owner, 2026-07-03, after observing the
3x speed-up). FSessionVideo.Fps int32 -> double.

+++ Source/AnomalyCapture/Private/AnomalyLabelWriter.h
@@ -86,7 +86,7 @@ namespace AnomalyLabel
 		FString VideoPath;           // "Video_Clip/<session>.mp4" (host ffmpeg produces it, Stage 3)
 		int32 ResolutionW = 0;
 		int32 ResolutionH = 0;
-		int32 Fps = 30;              // playback/encode metadata, NOT the capture throughput
+		double Fps = 30.0;           // MEASURED session rate (world-time; may be fractional) = mp4 playback fps
 		int32 TotalFrames = 0;
 	};
